pptx_merger/requirements.txt
python-pptx>=0.6.21
PyYAML>=6.0
lxml>=4.9.1
